requirements.txt
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
onesignal-sdk==2.0.0
packaging==24.2
sniffio==1.3.1
sqlparse==0.5.3
